Struct rds::array::csv::CSVFile

pub struct CSVFile {
    pub header: bool,
    pub flexible: bool,
    pub delimiter: u8,
    pub quote: u8,
    // some fields omitted
}

Structure representing a Comma Separated Value file.

Fields

header

A boolean indicating whether the first row should be treated as a header.

flexible

A boolean indicating whether a variable number of column should be allowed between rows.

delimiter

The value delimiter, b',' by default.

quote

The quote delimiter, b'"' by default.

Methods

impl CSVFile

fn new(path: &str) -> CSVFile

Allocate a new CSVFile structure with a given path. This function neither create nor open the file specified by the path.

fn read_row<T: FromStr + Clone>(&self, row_idx: usize) -> Result<NDArray<T>, String>

Open the CSV file for reading and read the row at row_idx (index starting at 0) as a one dimensional array. In case of failure, returns the error as a string.

fn read_column<T: FromStr + Clone>(&self, column_idx: usize) -> Result<NDArray<T>, String>

Open the CSV file for reading and read the column at column_idx (index starting at 0) as a one dimensional array. In case of failure, returns the error as a string.

fn read_2darray<T: FromStr + Clone>(&self) -> Result<NDArray<T>, String>

Open the CSV file for reading and read the whole file as a two dimensional array. In case of failure, returns the error as a string.

fn write_data<T: Display>(&mut self, data: &NDData<T>) -> Result<(), String>

Open (or create) the CSV file for writing and write a two dimensional array in it. This function overwrite any data already present in the file. In case of failure, returns the error as a string.