Struct rds::array::NDArray

pub struct NDArray<T> {
    // some fields omitted
}

Structure representing an owned n-dimensional array. The underlying storage is in row-major order.

Methods

impl<T: Clone> NDArray<T>

fn new(shape: &[usize], v: T) -> NDArray<T>

Allocate a new array of the specified shape with all elements initialized with the value v.

fn copy<R: NDData<T>>(data: &R) -> NDArray<T>

Allocate a new array which is a copy of data.

fn from_slice(shape: &[usize], data: &[T]) -> NDArray<T>

Allocate a new array from a row-major order contiguous array and a shape. The size of the shape (product of all its elements) must be equal to the array length.

fn reshape(&mut self, new_shape: &[usize])

Reshape an NDArray. The size of the new shape (product of all its elements) must be equal to the size of the current shape.