Trait rds::array::NDDataMut

pub trait NDDataMut<T: Clone + Display>: NDData<T> {
    fn get_data_mut(&mut self) -> &mut [T];

    fn idx_mut<'a>(&'a mut self, idx: &[usize]) -> &'a mut T { ... }
    fn transpose(&mut self) { ... }
    fn assign(&mut self, other: &NDData<T>) { ... }
}

A trait for struture giving mutable access to a N-dimensional array of type T.

Required Methods

fn get_data_mut(&mut self) -> &mut [T]

Return the underlying storage array as a mutable slice.

Provided Methods

fn idx_mut<'a>(&'a mut self, idx: &[usize]) -> &'a mut T

Take a slice of length N representing an -index in the array and return a mutable reference to the element at this position.

fn transpose(&mut self)

Perform a generic transpose. All the dimensions need to be the same.

fn assign(&mut self, other: &NDData<T>)

Assign another NDData to the NDDataMut. The shapes need to be identical.