Trait rds::array::NDSliceableMut

pub trait NDSliceableMut<'a, T: 'a>: NDSliceable<'a, T> {
    fn slice_mut(&'a mut self, idx: &[usize]) -> NDSliceMut<'a, T>;
}

A trait for N-dimensional data which can be sliced into a mutable sub slice.

Required Methods

fn slice_mut(&'a mut self, idx: &[usize]) -> NDSliceMut<'a, T>

Take a slice of length < N representing the sub slice index and return immutable borrow of the sub slice as an NDSlice. Because the storage is in row-major order and the slice need to be contiguous in the underlying storage array it means, for example, only the rows of a matrix can be borrowed.