Trait rds::blas::Blas

pub trait Blas<T: Clone + Display>: NDDataMut<T> {
    fn asum(&self) -> T;
    fn nrm2(&self) -> T;
    fn scal(&mut self, a: T);
    fn axpy(&mut self, a: T, x: &NDData<T>);
    fn dot(&self, x: &NDData<T>) -> T;
    fn gemv(&mut self, alpha: T, a: &NDData<T>, x: &NDData<T>, beta: T);
    fn gemm(&mut self, alpha: T, a: &NDData<T>, b: &NDData<T>, beta: T);
}

A trait representing N-dimensional array on which blas functions can be applied.

Required Methods

fn asum(&self) -> T

Compute an element-wise sum of the N-dimensional array.

fn nrm2(&self) -> T

Compute the euclidian norm of the N-dimensional array.

fn scal(&mut self, a: T)

Scale the N-dimensional array by a factor a.

fn axpy(&mut self, a: T, x: &NDData<T>)

Compute y += a * x where y is this array, x is a N-dimensional array of the same size as y and a is a scalar.

fn dot(&self, x: &NDData<T>) -> T

Compute the dot product of this array and x.

fn gemv(&mut self, alpha: T, a: &NDData<T>, x: &NDData<T>, beta: T)

Compute y = alpha * a * x + beta * y where y is this array, x is a one dimensional array, a a two dimensional array and alpha and beta are scalars. Automatically determine whether a need to be transposed.

fn gemm(&mut self, alpha: T, a: &NDData<T>, b: &NDData<T>, beta: T)

Compute y = alpha * a * x + beta * y where y is this array, a and b are two dimensional arrays and alpha and beta are scalars. Automatically determine whether a and b need to be transposed.